Initial react setup
