+++ package.json
@@ -51,6 +51,8 @@
     "passport-google-oauth20": "^2.0.0",
     "pug": "^3.0.2",
     "quill": "^1.3.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rsync": "^0.6.1",
     "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
