+++ public/assets/ts/admin/page-editor/EditorV2.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const EditorV2: React.FC<{
+  source: string;
+}> = ({ source }) => {
+  return <div>put an editor here for {source}</div>;
+};
