+++ tsconfig.json
@@ -3,6 +3,7 @@
     "declaration": false,
     "esModuleInterop": true,
     "incremental": true,
+    "jsx": "react",
     "module": "CommonJS",
     "outDir": "public/assets/js",
     "moduleResolution": "node",
@@ -10,7 +11,7 @@
     "resolveJsonModule": true,
     "sourceMap": false
   },
-  "include": ["public/assets/ts/**/*.ts"],
+  "include": ["public/assets/ts/**/*.ts", "public/assets/ts/admin/editor-loader.ts", "public/assets/ts/admin/generic-editor.tsx"],
   "ts-node": {
     "swc": true
   }
