+++ views/pages/admin/generic-editor.pug
@@ -29,7 +29,6 @@ block body
       +menuLayer(editors)
       button(id="close-side-menu-button") Hide
     div(id="body")
-      h2(id="editor-title")=editorName
       div(id="jsoneditor" data-source=editorSource)
       div(class="open-editor-container")
         button(class="pink-button disabled" id="open-editor-btn" disabled="true") Open Markdown Editor
